requirements.txt
IOPaint
torch
torchvision
opencv-python-headless
numpy
pillow